+++ tests/native-test.js
@@ -12,6 +12,16 @@ module.exports = async function ({
   setOverlay,
   store,
 }) {
+  if (process.env.TABLELIGHT_TEST_SCENARIO === 'concentration-reminder')
+    return require('./concentration-reminder-native')({
+      app,
+      controller,
+      getOverlay,
+      getState,
+      screen,
+      setOverlay,
+      store,
+    });
   if (process.env.TABLELIGHT_TEST_SCENARIO === 'concentration')
     return require('./concentration-native')({
       app,